# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RouteWeather API",
    "version": "1.0.0",
    "description": "Computes a route (Google Directions), samples points along the path, fetches weather forecasts (Tomorrow.io) at each estimated segment, and returns a timeline with a trip risk analysis."
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Local" }
  ],
  "paths": {
...
    "/v1/route-weather/plan/status/{jobId}": {
      "get": {
        "summary": "Poll the result of an enqueued plan",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job_id": { "type": "string" },
                    "status": { "type": "string", "enum": ["pending", "completed", "failed"] },
                    "result": { "$ref": "#/components/schemas/PlanResponse" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Structured error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Weather": {
        "type": "object",
        "properties": {
          "temperature_c": { "type": "number", "nullable": true },
          "rain_probability": { "type": "number", "nullable": true },
          "condition": { "type": "string", "nullable": true },
          "weather_code": { "type": "integer", "nullable": true },
          "precipitation_intensity_mm_h": { "type": "number", "nullable": true },
          "wind_speed_kmh": { "type": "number", "nullable": true },
          "visibility_km": { "type": "number", "nullable": true },
          "cloud_cover": { "type": "number", "nullable": true }
        }
      },
      "TimelineItem": {
        "type": "object",
        "properties": {
          "order": { "type": "integer" },
          "estimated_at": { "type": "string", "format": "date-time" },
          "eta_offset_seconds": { "type": "integer" },
          "distance_from_start_km": { "type": "number" },
          "location": {
            "type": "object",
            "properties": {
              "lat": { "type": "number" },
              "lng": { "type": "number" },
              "city": { "type": "string", "nullable": true }
            }
          },
          "weather": { "$ref": "#/components/schemas/Weather" }
        }
      },
      "PlanResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "properties": {
              "generated_at": { "type": "string", "format": "date-time" },
              "departure_at": { "type": "string", "format": "date-time" },
              "sample_interval_km": { "type": "number" },
              "warnings": { "type": "array", "items": { "type": "string" } }
            }
          },
          "route": {
            "type": "object",
            "properties": {
              "summary": { "type": "string" },
              "polyline": { "type": "string" },
              "total_distance_m": { "type": "number" },
              "total_duration_s": { "type": "integer" }
            }
          },
          "timeline": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TimelineItem" }
          },
          "risk": {
            "type": "object",
            "properties": {
              "score": { "type": "integer" },
              "alerts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "type": { "type": "string" },
                    "label": { "type": "string" },
                    "severity": { "type": "string" },
                    "segment": { "type": "string" },
                    "order": { "type": "integer" }
                  }
                }
              },
              "summary": { "type": "string" }
            }
          }
        }
      }
    }
  }
}